modal enabled, navbar with conditional rendering
