+++ src/js/store/flux.js
@@ -12,7 +12,8 @@ const getState = ({ getStore, getActions, setStore }) => {
 					background: "white",
 					initial: "white"
 				}
-			]
+			],
+			isLogged: true
 		},
 		actions: {
 			// Use getActions to call a function within a fuction
